Add git log to git commands script
Also:
* Improve organization of git commands script
* Improve documentation of git log script
